refactor(analytics): decompose CodebaseAnalytics.vue script into composables (#2228, #2230)
Extract ~2,600 lines from the script section into 4 focused composables:

- useSourceRegistry.ts (220 lines) — source CRUD, selection, knowledge base
- useAnalyticsDataFetchers.ts (1,144 lines) — stats, problems, charts, deps
- useCodeIntelAnalysis.ts (1,923 lines) — security/perf/redis scores, smells
- useAnalyticsDebug.ts (303 lines) — NPU test, endpoint tests, utilities

Parent script reduced from ~2,734 to ~370 lines of pure orchestration
(composable wiring, lifecycle hooks, event routing).

Total file: 4,148 → 1,928 lines (54% reduction).
